feat(aztec.js): implement proof versioning dir structure
aztec.js proof construction versioning to follow same uint24 _proof structure as protocol package
