+++ packages/aztec.js/src/proof/epoch0/BURN/burn/verifier.js
@@ -1,0 +1,5 @@
+const JoinSplitFluidVerifier = require('../../../verifier');
+
+class BurnVerifier66305 extends JoinSplitFluidVerifier {}
+
+module.exports = BurnVerifier66305;
