+++ packages/aztec.js/src/proof/epoch0/MINT/mint/verifier.js
@@ -1,0 +1,5 @@
+const JoinSplitFluidVerifier = require('../../../verifier');
+
+class MintVerifier66049 extends JoinSplitFluidVerifier {}
+
+module.exports = MintVerifier66049;
