+++ packages/aztec.js/src/proof/epoch0/UTILITY/publicRange/helpers.js
