--- packages/aztec.js/src/proof/joinSplitFluid/burn/verifier.js
@@ -1,5 +1,0 @@
-const JoinSplitFluidVerifier = require('../verifier');
-
-class BurnVerifier extends JoinSplitFluidVerifier {}
-
-module.exports = BurnVerifier;
