--- packages/aztec.js/src/proof/joinSplitFluid/mint/verifier.js
@@ -1,5 +1,0 @@
-const JoinSplitFluidVerifier = require('../verifier');
-
-class MintVerifier extends JoinSplitFluidVerifier {}
-
-module.exports = MintVerifier;
