+++ packages/aztec.js/test/proof/joinSplit/verifier.js
@@ -8,7 +8,7 @@ const { padLeft, randomHex } = require('web3-utils');
 
 const { balancedPublicValues, mockNoteSet, randomNoteValue } = require('../../helpers/note');
 const { JoinSplitProof, Proof } = require('../../../src/proof');
-const JoinSplitVerifier = require('../../../src/proof/joinSplit/verifier');
+const JoinSplitVerifier = require('../../../src/proof/epoch0/BALANCED/joinSplit/verifier');
 const { mockZeroJoinSplitProof } = require('../../helpers/proof');
 const ProofUtils = require('../../../src/proof/utils');
 
@@ -56,7 +56,6 @@ describe('Join-Split Proof Verifier', () => {
 
             const verifier = new JoinSplitVerifier(proof);
             const result = verifier.verifyProof();
-            // console.log(verifier.errors);
             expect(verifier.isValid).to.equal(true);
             expect(result.pairingGammas).to.equal(undefined);
             expect(result.pairingSigmas).to.equal(undefined);
